Rewrite the XML part
This patch rewrites the XML part completely to use libxml2. This fix all
issue with invalid XML output.

With this patch there might be dropped trackpoints which put to invalide
XML before. This trackpoins was encountered as timestamp only in the
case when the timestamp is one second before the activity start.

See also http://developer.garmin.com/schemas/tcx/v2/ for schema
documentation.