AntFr310XT: bugfix: improve stopAsync()
Sometimes Ctrl-C was not effective. As the state change in stopAsync() was overwritten by a returning listener afterwards. Fix it, by checking for this in changeState().

This is still not perfect, as setting ST_ANTFS_LAST might not be enough for stopping immediately, as other thread might be sleeping in a listener, and we stop only when that returns.

Signed-off-by: RALOVICH, Kristóf <tade60@freemail.hu>